refactor(web): move all types into typings folder
