+++ web/src/features/dialog/components/fields/date.tsx
@@ -1,4 +1,4 @@
-import { IDateInput } from '../../../../interfaces/dialog';
+import { IDateInput } from '../../../../typings/dialog';
 import { Control, useController } from 'react-hook-form';
 import { FormValues } from '../../InputDialog';
 import { DatePicker, DateRangePicker, TimeInput } from '@mantine/dates';
