+++ web/src/features/dialog/components/fields/input.tsx
@@ -1,7 +1,7 @@
 import { createStyles, PasswordInput, TextInput } from '@mantine/core';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React from 'react';
-import { IInput } from '../../../../interfaces/dialog';
+import { IInput } from '../../../../typings/dialog';
 import { UseFormRegisterReturn } from 'react-hook-form';
 
 interface Props {
