+++ web/src/features/progress/Progressbar.tsx
@@ -3,11 +3,7 @@ import { Box, Text, createStyles } from '@mantine/core';
 import { useNuiEvent } from '../../hooks/useNuiEvent';
 import { fetchNui } from '../../utils/fetchNui';
 import ScaleFade from '../../transitions/ScaleFade';
-
-export interface ProgressbarProps {
-  label: string;
-  duration: number;
-}
+import type { ProgressbarProps } from '../../typings';
 
 const useStyles = createStyles((theme) => ({
   container: {
