+++ web/src/typings/alert.ts
@@ -1,0 +1,12 @@
+export interface AlertProps {
+  header: string;
+  content: string;
+  centered?: boolean;
+  size?: 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+  overflow?: boolean;
+  cancel?: boolean;
+  labels?: {
+    cancel?: string;
+    confirm?: string;
+  };
+}
