+++ web/src/typings/dialog.ts
@@ -1,5 +1,13 @@
 import { IconProp } from '@fortawesome/fontawesome-svg-core';
 
+export interface InputProps {
+  heading: string;
+  rows: Array<IInput | ICheckbox | ISelect | INumber | ISlider | IColorInput | IDateInput | ITextarea | ITimeInput>;
+  options?: {
+    allowCancel?: boolean;
+  };
+}
+
 type BaseField<T, U> = {
   type: T;
   label: string;
