+++ web/src/typings/progress.ts
@@ -1,0 +1,11 @@
+export interface CircleProgressbarProps {
+  label?: string;
+  duration: number;
+  position?: 'middle' | 'bottom';
+  percent?: boolean;
+}
+
+export interface ProgressbarProps {
+  label: string;
+  duration: number;
+}
