+++ web/src/typings/radial.ts
@@ -1,0 +1,6 @@
+import { IconProp } from '@fortawesome/fontawesome-svg-core';
+
+export interface RadialMenuItem {
+  icon: IconProp;
+  label: string;
+}
