+++ web/vite.config.ts
@@ -1,5 +1,6 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+const path = require("path");
 
 // https://vitejs.dev/config/
 export default defineConfig({
